Feat(CLI syntax): moved from old to new CLI syntax
Changed the taught syntax to follow the one introduced in version 1.13. All the commands relating to managing containers were moved to `docker container` and images were moved to `docker image` (plus some misc commands). Included a callout briefly explaining the differences and purpose of the new syntax.

Included a new output for the `docker image build --help` command to the new(er) version.